requirements_dev.txt
check-jsonschema==0.33.3
PyQt5==5.15.11
pytest-qt==4.5.0
pytest-github-actions-annotate-failures==0.3.0